+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Braintree iOS SDK Release Notes
 
-## unreleased
+## 6.25.0 (2024-12-11)
 * BraintreePayPal
   * Add `shippingCallbackURL` to `BTPayPalCheckoutRequest`
   * Add `BTPayPalRequest.userPhoneNumber` optional property
+  * Send `url` in `event_params` for App Switch events to PayPal's analytics service (FPTI)
 * BraintreeVenmo
   * Send `url` in `event_params` for App Switch events to PayPal's analytics service (FPTI)
+  * Add `BTVenmoClient(apiClient:universalLink:)` to use Universal Links when redirecting back from the Venmo flow
+* BraintreeCore
+  * Deprecate `BTAppContextSwitcher.sharedInstance.returnURLScheme`
+* BraintreeThreeDSecure
+  * Add `BTThreeDSecureRequest.requestorAppURL`
 
 ## 6.24.0 (2024-10-15)
 * BraintreePayPal